DELIA-59648 [Coverity] Various medium impact issues identified in aamp - part9
Reason for change: Resolved part-9 coverity issue(Odd Sprint 1802 V3) - 121357,335450,335448,330393,329356,329354,329353,186827
Test Procedure: general regression
Risks: Low

(cherry picked from commit d8d490070d1e01c921348d8087447d47188187fd)

+++ isobmff/isobmffbox.h
@@ -32,7 +32,7 @@
 #include "AampLogManager.h"
 
 #define READ_U32(buf) \
-	(buf[0] << 24) | (buf[1] << 16) | (buf[2] << 8) | buf[3]; buf+=4;
+	((uint32_t)buf[0] << 24) | ((uint32_t)buf[1] << 16) | ((uint32_t)buf[2] << 8) | buf[3]; buf+=4;
 
 #define WRITE_U64(buf, val) \
 	buf[0]= val>>56; buf[1]= val>>48; buf[2]= val>>40; buf[3]= val>>32; buf[4]= val>>24; buf[5]= val>>16; buf[6]= val>>8; buf[7]= val;
